Moved Pathfinders component to club tab.
